Add ability to use pre-created keys
This will allow us to turn off the creation of the Vaults in the LZ,
and use pre-existing keys for our logging buckets.

You can turn off the vault creation by setting the vault_type(s) to
be "NONE"

The existing keys will need to have a policy set allowing the
objectstorage service and a few others to use the keys.

An example of the needed policy, assuming the external keys are in
a compartment named "Keybox" and the LZ is deployed in us-ashburn-1
region:

  Allow service objectstorage-us-ashburn-1 to use keys in compartment Keybox
  Allow service blockstorage,FssOc1Prod, OKE, streaming to use keys in compartment Keybox